ath9k: only enable MSI on AR9565
We have multiple user reports of AR9462 having difficulty connecting
to wireless networks, and we believe that this problem occurs since
we switched to MSI interrupts.

In the Taipei office, we reproduced an issue where Bobby's laptop
will be disconnected from the AP before it completes a 1000 second
ping test. This was bisected down to the MSI support commit.

Our AR9565 platforms pass the same test with MSI interrupts enabled,
so resrict the MSI interrupt support to AR9565-only, where we have
confirmed that it's working fine.

AR9462 and other hardware revisions will now go back to legacy
interrupts.

https://phabricator.endlessm.com/T19774