tpm: fix suspend failure on Acer Veriton X4110G
The TPM on Acer Veriton X4110G shows the following message in the
first resume.
  tpm tpm0: A TPM error(38) occured continue selftest
And error(38) is TPM_ERR_INVALID_POSTINIT which means the TPM is
not in the correct state. Then the following suspend will always
fail with the following messages.
  tpm tpm0: Error (38) sending savestate before suspend
  PM: Device 00:0b failed to suspend: error 38

This commit log the unexpected TPM state and skip the TPM_SaveState
in following suspends.

https://phabricator.endlessm.com/T20438

Signed-off-by: Chris Chiu <chiu@endlessm.com>